+++ apps/elf-api/tests/http/request_validation/health.rs
@@ -1,0 +1,33 @@
+use axum::{
+	body::Body,
+	http::{Request, StatusCode},
+};
+use tower::util::ServiceExt as _;
+
+use crate::helpers;
+use elf_api::{routes, state::AppState};
+
+#[tokio::test]
+#[ignore = "Requires external Postgres and Qdrant. Set ELF_PG_DSN and ELF_QDRANT_GRPC_URL (or ELF_QDRANT_URL) to run."]
+async fn health_ok() {
+	let Some((test_db, qdrant_url, collection)) = helpers::test_env().await else {
+		return;
+	};
+	let config = helpers::test_config(test_db.dsn().to_string(), qdrant_url, collection);
+	let state = AppState::new(config).await.expect("Failed to initialize app state.");
+	let app = routes::router(state.clone());
+	let _ = routes::admin_router(state);
+	let response = app
+		.oneshot(
+			Request::builder()
+				.uri("/health")
+				.body(Body::empty())
+				.expect("Failed to build request."),
+		)
+		.await
+		.expect("Failed to call /health.");
+
+	assert_eq!(response.status(), StatusCode::OK);
+
+	test_db.cleanup().await.expect("Failed to cleanup test database.");
+}
